Stats "déjà joué", paliers décimaux et nettoyage docs (#26)
## Ce qui change

### Stats sur l'écran "déjà joué"
- Nouvel endpoint `GET /api/stats/today` : score du joueur, médiane des joueurs (PostgreSQL `PERCENTILE_CONT(0.5)`), taux d'échec et moyenne d'écoute par morceau
- Écran "déjà joué" redessiné : ton score vs "Aujourd'hui les joueurs font X pts", accordéon de détail par morceau (pochette Deezer + lien), compte à rebours jusqu'à minuit UTC dans le sous-titre
- Le bloc score se masque automatiquement si l'écran est trop petit (pas de scroll)

### Paliers d'écoute décimaux
- Paliers passent de `[1, 2, 3, 5, 10, 15, 30]` à `[0.5, 1, 1.5, 2, 3, 5, 10]` secondes
- `ListenedDurationSeconds` et `TotalDurationSeconds` : `int` → `decimal` (back + migration EF)
- `DurationScores` : `Dictionary<int,int>` → `Dictionary<decimal,int>`, nouveaux scores `1000/850/700/550/400/250/100`
- Migration `DecimalDurations` : alters colonnes + `migrationBuilder.Sql(UPDATE ...)` pour mettre à jour les Settings existants en prod (contrairement à `UpdateData` qui ne s'applique pas sur une DB existante)

### Docs
- `BACKEND_STRUCTURE_FR.md` réécrit (supprimé SQL Server, services obsolètes, ancien ScoreCalculator)
- `TACHES.md` mis à jour : Leaderboard et Auth Register classés en décisions définitives abandonnées
- `CLAUDE.md`, `COMMENCE_ICI_FR.md`, `FRONTEND_STRUCTURE_FR.md` : paliers et état projet mis à jour

+++ src/back/InSeconds.Api.UnitTests/Common/Settings/AppSettingsBindingTests.cs
@@ -34,7 +34,7 @@ public sealed class AppSettingsBindingTests
         result.AllowedDurationsSeconds.Should().Equal(1, 3, 5);
         result.MaxExtensionsPerAnswer.Should().Be(2);
         result.TracksPerChallenge.Should().Be(5);
-        result.DurationScores.Should().BeEquivalentTo(new Dictionary<int, int> { [1] = 900, [3] = 600, [5] = 400 });
+        result.DurationScores.Should().BeEquivalentTo(new Dictionary<decimal, int> { [1m] = 900, [3m] = 600, [5m] = 400 });
     }
 
     [Fact]
