+++ src/back/InSeconds.Api/Domain/GameSession.cs
@@ -6,7 +6,7 @@ public sealed class GameSession
     public Guid PlayerId { get; set; }
     public int DailyChallengeId { get; set; }
     public int TotalScore { get; set; }
-    public int TotalDurationSeconds { get; set; }
+    public decimal TotalDurationSeconds { get; set; }
     public DateTime CreatedAt { get; set; }
 
     public Player Player { get; set; } = null!;
